blink1control: fixed IFTTT event issue #154. It was bad datetime type, now using qint64 for secs-since-epoch

+++ qt/blink1control/mainwindow.h
@@ -327,7 +327,7 @@ private:
 
     QTimer *inputsTimer;
     int inputTimerCounter;
-    int lastIftttDate;    // timestamp of the last IFTTT event we received
+    qint64 lastIftttDate;    // timestamp of the last IFTTT event we received
     bool isIftttChecked;  // FIXME: only check IFTTT once, even tho multiple DataInputs for it
 
     HttpServer *httpserver;
